packages/aggregator/pyproject.toml
[project]
name = "aggregator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Andy Eschbacher", email = "andyesch@protonmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "functions-framework>=3.0.0",
    "google-cloud-storage>=3.3.0",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "pytz>=2025.2",
    "requests>=2.32.4",
]

[build-system]
requires = ["uv_build>=0.8.4,<0.9.0"]
build-backend = "uv_build"

[tool.setuptools.packages.find]
where = ["src"]
